requirements.txt
Flask
Flask-CORS
requests
beautifulsoup4
selenium
pandas
numpy
gunicorn
python-dotenv
